+++ lib/interceptor/telegram_error_interceptor.dart
@@ -15,22 +15,30 @@ class TelegramErrorInterceptor extends Interceptor {
 
   final List<int> ignoreStatusCodes;
 
+  /// Include headers in the error message
+  final bool includeHeaders;
+
   /// Singleton instance
   static TelegramErrorInterceptor? _instance;
 
   /// Singleton factory
-  factory TelegramErrorInterceptor(
-      {required String botToken,
-      required int chatId,
-      required List<int> ignoreStatusCodes}) {
-    _instance ??=
-        TelegramErrorInterceptor._internal(botToken, chatId, ignoreStatusCodes);
+  factory TelegramErrorInterceptor({
+    required String botToken,
+    required int chatId,
+    required List<int> ignoreStatusCodes,
+    required bool includeHeaders,
+  }) {
+    _instance ??= TelegramErrorInterceptor._internal(botToken, chatId, ignoreStatusCodes, includeHeaders);
     return _instance!;
   }
 
   /// Private constructor
   TelegramErrorInterceptor._internal(
-      this.botToken, this.chatId, this.ignoreStatusCodes);
+    this.botToken,
+    this.chatId,
+    this.ignoreStatusCodes,
+    this.includeHeaders,
+  );
 
   /// Send error message to Telegram function
   Future<void> sendErrorToTelegram(String errorMessage) async {
@@ -47,10 +55,10 @@ class TelegramErrorInterceptor extends Interceptor {
   }
 
   @override
-  Future<void> onError(
-      DioException err, ErrorInterceptorHandler handler) async {
+  Future<void> onError(DioException err, ErrorInterceptorHandler handler) async {
     String errorMessage;
     String sticker;
+    String requestHeaders = '';
 
     /// Get the request URL, status code, status message, and error message
     String url = escapeMarkdown(err.requestOptions.uri.toString());
@@ -58,61 +66,86 @@ class TelegramErrorInterceptor extends Interceptor {
     String deviceSticker = getDeviceSticker();
     String device = await getDevice();
 
+    if (includeHeaders) {
+      /// Get the request headers
+      err.requestOptions.headers.forEach((key, value) {
+        requestHeaders += "$key: $value\n";
+      });
+    }
+
     /// Define sticker and create an error message with stickers for each line
     switch (err.type) {
       case DioExceptionType.sendTimeout:
         sticker = '⏰';
-        errorMessage = "$sticker *Send Timeout Error*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Send Timeout Error*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
 
       case DioExceptionType.receiveTimeout:
         sticker = '⏳';
-        errorMessage = "$sticker *Receive Timeout Error*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Receive Timeout Error*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
 
       case DioExceptionType.cancel:
         sticker = '🚫';
-        errorMessage = "$sticker *Request Cancelled*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Request Cancelled*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
 
       case DioExceptionType.connectionTimeout:
         sticker = '🔗';
-        errorMessage = "$sticker *Connection Timeout*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Connection Timeout*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
 
       case DioExceptionType.badCertificate:
         sticker = '📜';
-        errorMessage = "$sticker *Bad Certificate Error*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Bad Certificate Error*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
 
       case DioExceptionType.connectionError:
         sticker = '🔌';
-        errorMessage = "$sticker *Connection Error*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Connection Error*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
+      case DioExceptionType.badResponse:
+        sticker = '🔌';
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Bad Response*\n\n"
+            "$deviceSticker *Device:* #$device\n"
+            "🔴 *Method:* `${err.requestOptions.method}`\n"
+            "⚠️ *Status Code:* `${err.response?.statusCode}`\n"
+            "🌐 *URL:* `$url`\n"
+            "${includeHeaders ? "📥 *Request Headers:*\n$requestHeaders\n" : ''}"
+            "📝 *Request Data:* ${err.requestOptions.data?.toString() ?? 'No request data'}\n"
+            "📄 *Response Data:* ${err.response?.data?.toString() ?? 'No response data'}";
+        break;
 
-      case DioExceptionType.unknown:
       default:
         sticker = '🤷🏻‍♀️🤷🏻‍♂️';
-        errorMessage = "$sticker *Unknown Error*\n\n"
-            "$deviceSticker *Device:* $device\n"
+        errorMessage = "#️⃣TAGS: #${err.requestOptions.method}, #${err.response?.statusCode}, #${err.type.name}\n"
+            "$sticker *Unknown Error*\n\n"
+            "$deviceSticker *Device:* #$device\n"
             "💬 *Message:* $errMessage\n"
             "🌐 *URL:* `$url`";
         break;
@@ -126,27 +159,30 @@ class TelegramErrorInterceptor extends Interceptor {
   }
 
   @override
-  Future<void> onResponse(
-      Response response, ResponseInterceptorHandler handler) async {
-    if (((response.statusCode ?? 0) < 200 ||
-            (response.statusCode ?? 0) >= 300) &&
+  Future<void> onResponse(Response response, ResponseInterceptorHandler handler) async {
+    if (((response.statusCode ?? 0) < 200 || (response.statusCode ?? 0) >= 300) &&
         !ignoreStatusCodes.contains(response.statusCode)) {
       String sticker = '🔴';
       String method = escapeMarkdown(response.requestOptions.method);
       String url = escapeMarkdown(response.requestOptions.uri.toString());
       String statusCode = escapeMarkdown(response.statusCode.toString());
-      String requestMessage = escapeMarkdown(
-          response.requestOptions.data?.toString() ?? 'No request data');
-      String responseData =
-          escapeMarkdown(response.data?.toString() ?? 'No response data');
+      String requestMessage = escapeMarkdown(response.requestOptions.data?.toString() ?? 'No request data');
+      String responseData = escapeMarkdown(response.data?.toString() ?? 'No response data');
       String deviceSticker = getDeviceSticker();
       String device = await getDevice();
+      String requestHeaders = '';
 
-      String errorMessage = "$sticker *Bad Response*\n\n"
-          "$deviceSticker *Device:* $device\n"
-          "🔴 *Method:* `$method`\n"
-          "⚠️ *Status Code:* `$statusCode`\n"
+      if (includeHeaders) {
+        response.requestOptions.headers.forEach((key, value) {
+          requestHeaders += "$key: $value\n";
+        });
+      }
+
+      String errorMessage = "#️⃣TAGS: #$method, #$statusCode\n"
+          "$sticker *Bad Response*\n\n"
+          "$deviceSticker *Device:* #$device\n"
           "🌐 *URL:* `$url`\n"
+          "${includeHeaders ? "📥 *Request Headers:*\n$requestHeaders\n" : ''}"
           "📝 *Request Data:* $requestMessage\n"
           "📄 *Response Data:* $responseData";
       sendErrorToTelegram(errorMessage);
@@ -157,8 +193,7 @@ class TelegramErrorInterceptor extends Interceptor {
   /// Escape MarkdownV2 special characters
   String escapeMarkdown(String text) {
     return text.replaceAllMapped(
-        RegExp(r'([_*`$begin:math:display$$end:math:display${}()~>#+\-=|.!])'),
-        (match) => '\\${match[0]}');
+        RegExp(r'([_*`$begin:math:display$$end:math:display${}()~>#+\-=|.!])'), (match) => '\\${match[0]}');
   }
 
   Future<String> getDevice() async {