+++ lib/telegram_bot_crashlytics.dart
@@ -11,6 +11,7 @@ class TelegramBotCrashlytics {
   final int chatId;
 
   final List<int> ignoreStatusCodes;
+  final bool includeHeaders;
 
   static TelegramBotCrashlytics? _instance;
 
@@ -18,21 +19,28 @@ class TelegramBotCrashlytics {
     required this.botToken,
     required this.chatId,
     required this.ignoreStatusCodes,
+    required this.includeHeaders,
   }) {
     _telegramErrorInterceptor = TelegramErrorInterceptor(
-        botToken: botToken,
-        chatId: chatId,
-        ignoreStatusCodes: ignoreStatusCodes);
+      botToken: botToken,
+      chatId: chatId,
+      ignoreStatusCodes: ignoreStatusCodes,
+      includeHeaders: includeHeaders,
+    );
   }
 
-  factory TelegramBotCrashlytics(
-      {required String botToken,
-      required int chatId,
-      List<int>? ignoreStatusCodes}) {
+  factory TelegramBotCrashlytics({
+    required String botToken,
+    required int chatId,
+    List<int>? ignoreStatusCodes,
+    bool? includeHeaders,
+  }) {
     _instance ??= TelegramBotCrashlytics._internal(
-        botToken: botToken,
-        chatId: chatId,
-        ignoreStatusCodes: ignoreStatusCodes ?? []);
+      botToken: botToken,
+      chatId: chatId,
+      ignoreStatusCodes: ignoreStatusCodes ?? [],
+      includeHeaders: includeHeaders ?? false,
+    );
     return _instance!;
   }
 
@@ -42,8 +50,7 @@ class TelegramBotCrashlytics {
 
   static TelegramBotCrashlytics get instance {
     if (_instance == null) {
-      throw Exception(
-          "TelegramBotCrashlytics instance not initialized. Please call the constructor first.");
+      throw Exception("TelegramBotCrashlytics instance not initialized. Please call the constructor first.");
     }
     return _instance!;
   }
@@ -51,13 +58,12 @@ class TelegramBotCrashlytics {
   /// Send error message to Telegram function
   Future<void> sendErrorToTelegram(String errorMessage) async {
     if (errorMessage.isEmpty) return;
-    await _telegramErrorInterceptor.sendErrorToTelegram(
-        "🚨 *Error occurred in the application*\n\n📝 *Message:* _${errorMessage}_");
+    await _telegramErrorInterceptor
+        .sendErrorToTelegram("🚨 *Error occurred in the application*\n\n📝 *Message:* _${errorMessage}_");
   }
 
   Future<void> sendInfoToTelegram(String message) async {
     if (message.isEmpty) return;
-    await _telegramErrorInterceptor
-        .sendErrorToTelegram("📢 *Information*\n\n📝 *Message:* _${message}_");
+    await _telegramErrorInterceptor.sendErrorToTelegram("📢 *Information*\n\n📝 *Message:* _${message}_");
   }
 }